ci: Label Pull Request (#28)
# Pull Request

## Description

This pull request introduces a new automated labeling system for pull
requests, along with a workflow to add size labels based on the number
of lines changed. The key changes are as follows:

Labeling Configuration:

*
[`.github/other-configurations/labeller.yml`](diffhunk://#diff-b490f987d9b055ce6e06e37d13763d61f6e0d5ee349b390bbb910f8d88ff36b9R1-R44):
Added a configuration file to define labels based on file changes and
branch names. This includes labels for documentation, Rust, Python,
markdown, dependencies, Justfile, shell scripts, GitHub actions, and git
hooks.

Workflow Updates:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR19-R44):
Added a new job to the workflow for labeling pull requests. This job
uses the `actions/labeler` action to apply labels based on the
configuration file and the `pascalgn/size-label-action` to add size
labels based on the number of lines changed.

Fixes #16