chore: Init Rust Project (#31)
This pull request introduces a new Rust project named `depup` and
includes the initial setup for the project. The most important changes
are the addition of the `Cargo.toml` file and the creation of the main
Rust source file.

Project setup:

*
[`Cargo.toml`](diffhunk://#diff-2e9d962a08321605940b5a657135052fbcef87b5e360662bb527c96d9a615542R1-R6):
Added the package configuration for the new Rust project, specifying the
name, version, and edition.

Initial code:

*
[`src/main.rs`](diffhunk://#diff-42cb6807ad74b3e201c5a7ca98b911c5fa08380e942be6e4ac5807f8377f87fcR1-R3):
Created the main source file with a simple `main` function that prints
"Hello, world!" to the console.

Fixes #23