ci: Add Dependency Review (#32)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for dependency review. The most important change is the addition of a
`dependency-review` job to the
`.github/workflows/pull-request-tasks.yml` file.

### Changes:

*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfR45-R55):
Added a new `dependency-review` job that uses the
`actions/checkout@v4.2.2` and `actions/dependency-review-action@v4.5.0`
actions to perform a dependency review on the repository.

Fixes #19